+++ packages/desktop_auth0_flutter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0
+
+- Expose optional AppLinks instance in DesktopAuth0FlutterInitOptions
+- Expose waitUrlTimeoutBuilder in DesktopAuth0FlutterInitOptions.
+  This allows users of the library to change the "wait for url" timeout logic.
+
 ## 0.1.2
 
 - BUG: Fix minTTL logic